Start of https://github.com/Ree6-Applications/Ree6/issues/125
